改写一些样式和帐号展品功能的完成

+++ src/components/account/Admin.vue
@@ -1,6 +1,5 @@
 <template>
   <div>
-    <el-button type="primary" @click="currentChange">刷新</el-button>
     <el-button class="add-btn" type="primary" @click="addAdmin">添加管理员</el-button>
     <el-table style="width: 100%" :data="adminList">
       <el-table-column label="管理员帐号" prop="userName" width="150"></el-table-column>
@@ -9,7 +8,6 @@
       <el-table-column label="账号类型" prop="memberAccountTypeId" width="150"></el-table-column>
       <el-table-column label="开户日期" prop="openDate" width="200"></el-table-column>
       <el-table-column label="最后登录时间" prop="lastLoginDate" width="200"></el-table-column>
-      <el-table-column v-if="false" prop="id"></el-table-column>
       <el-table-column label="编辑" fixed="right" width="150">
         <template slot-scope="scope">
           <el-button size="mini" @click="handleEdit(scope.$index, scope.row)">修改</el-button>
@@ -30,7 +28,12 @@
         </div>
       </el-col>
     </el-row>
-    <admin-edit :editDialog.sync="editDialog" :id="id" :curpage='paginations.currentPage' @currentChange="currentChange"></admin-edit>
+    <admin-edit
+      :editDialog.sync="editDialog"
+      :editAdmin.sync="editAdmin"
+      :curpage="paginations.currentPage"
+      @currentChange="currentChange"
+    ></admin-edit>
     <add-admin :addDialog.sync="addDialog" @getAllAdminList="getAllAdminList"></add-admin>
   </div>
 </template>
@@ -39,26 +42,43 @@
 import AdminEdit from "./Edit";
 import AddAdmin from "./Add";
 import { getAdminInfo, deleteAdminInfo } from "../../api/adminList";
-import {handleAdminData} from '../../lib/handleData'
+import { handleAdminData } from "../../lib/handleData";
 export default {
   name: "Admin",
   methods: {
     handleEdit(index, row) {
-      this.id = row.id;
+      this.editAdmin = {
+        id: row.id,
+        password: null,
+        checkpass: null,
+        memberAccountTypeId: row.memberAccountId,
+        telephone: row.telephone
+      };
+      console.log(row);
       this.editDialog = true;
     },
     //删除对应的管理员信息
     handleDelete(index, row) {
-      let id = row.id;
-      deleteAdminInfo({ id }).then(result => {
-        if (result.data.status === 200) {
-          this.$message.success({
-            message: "删除成功",
-            duration: 1500
-          }),
-            this.currentChange(this.paginations.currentPage);
-        }
-      });
+      this.$confirm("此操作将永久删除,是否继续?", "提示", {
+        cancelButtonText: "取消",
+        confirmButtonText: "确定",
+        type: "warning"
+      })
+        .then(() => {
+          let id = row.id;
+          deleteAdminInfo({ id }).then(result => {
+            if (result.data.status === 200) {
+              this.$message.success("删除成功"),
+                this.currentChange(this.paginations.currentPage);
+            }
+          });
+        })
+        .catch(() => {
+          this.$message({
+            type: "info",
+            message: "已取消删除"
+          });
+        });
     },
     addAdmin() {
       this.addDialog = true;
@@ -70,33 +90,38 @@ export default {
         rows: 8
       }).then(result => {
         // handleData将获取的数据进行一些转换
-        if(result.data.status === 200){
-       this.adminList= handleAdminData(result)
-        this.paginations.total = result.data.info.total;
+        if (result.data.status === 200) {
+          this.adminList = handleAdminData(result);
+          this.paginations.total = result.data.info.total;
         }
         console.log(result);
       });
     },
     currentChange(val) {
-      this.paginations.currentPage=val
+      this.paginations.currentPage = val;
       getAdminInfo({
         page: this.paginations.currentPage,
         rows: 8
       }).then(result => {
-        if(result.data.status === 200){
-        this.adminList= handleAdminData(result)
+        if (result.data.status === 200) {
+          this.adminList = handleAdminData(result);
         }
         console.log(result);
       });
-    },
-
+    }
   },
   data() {
     return {
       adminList: [],
       editDialog: false,
       addDialog: false,
-      id: null,
+      editAdmin: {
+        id: null,
+        password: null,
+        checkpass: null,
+        memberAccountTypeId: "",
+        telephone: null
+      },
       paginations: {
         total: 0,
         currentPage: 1,
@@ -119,7 +144,7 @@ export default {
   float: right;
   margin-bottom: 20px;
 }
-.pagination{
+.pagination {
   text-align: center;
 }
 </style>
